backend/requirements.txt
fastapi[all]>=0.110.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.40
python-multipart>=0.0.20
python-dotenv>=1.0.1
anthropic>=0.62.0
httpx>=0.25.0
markdown>=3.5
pypandoc>=1.12
pydantic>=2.11.4
email-validator>=2.2.0
itsdangerous>=2.2.0
orjson>=3.10.18
pydantic-extra-types>=2.10.4
pydantic-settings>=2.9.1
ujson>=5.10.0
httptools>=0.6.4
uvloop>=0.21.0
watchfiles>=1.0.5
aiohttp>=3.8.0
